refactor: Service 레이어가 과연 필요한가?를 고민한 후 Service 클래스 삭제
